// tests/model/Applicant.ts
export interface Applicant {
    fullName: string;
    fatherName: string;
    motherName: string;
    birthDay: string;
    city: string;
    currentAddress: string;
    permanentAddress: string;
    mobileNumber: string;
    guardianMobileNumber: string;
    email: string;
    tribalStatus: boolean;
    childOfFreedonFighter: boolean;
    department: string;
    examName: string;
    passingYear: string;
    group: string;
    board: string;
    sscRoll: string;
    sscRegistration: string;
    gpa: string;
    avatar: string;
    marksheetAvatar?: string;
}